✏️ update : 고객 추가하기 확인 버튼 수정 & 그룹 더보기 선택 그룹 클릭 & id Long 변경
고객 추가하기 확인 버튼 수정 & 그룹 더보기 선택 그룹 클릭 & id Long 변경

> Related to : #74

+++ app/src/main/java/com/example/gajamap/data/model/ClientResponse.kt
@@ -35,5 +35,5 @@ data class PutClientRequest(
 )
 
 data class DeleteRequest(
-    val clientIds : List<Int>
+    val clientIds : List<Long>
 )
